rename dbtest to testutil
